requirements.txt
matplotlib~=3.1.3
numpy~=1.20.1
pybullet~=3.0.8
pandas~=1.2.3
torch~=1.8.1
lxml~=4.5.2

transform3d==0.0.3
# scipy==1.3.1  it has a dependency error
scipy==1.4.0 # instead
torchvision~=0.9.1
Pillow~=7.2.0
tqdm~=4.56.0
ipython~=7.20.0